pyproject.toml
[build-system]
requires = [
    "setuptools>=64.0",
    "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "SimplHDL"
dynamic = ["version"]
description = "A framework for simulating and implementing HDL designs"
readme = "README.md"
authors = [
    {name = "Rasmus Grøndahl Olsen", email = "simplhdl@gmail.com"},
]
maintainers = [
    {name = "Rasmus Grøndahl Olsen", email = "simplhdl@gmail.com"},
]
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "argcomplete",
    "black",
    "flake8",
    "GitPython",
    "importlib-resources;python_version<'3.9'",
    "jinja2",
    "matplotlib",
    "networkx",
    "peakrdl-html",
    "peakrdl-pyuvm",
    "peakrdl-regblock>=1.2.0",
    "pyyaml",
    "rich",
    "singleton-decorator",
    "systemrdl-compiler",
    "types-PyYAML",
    "vsg",
]

[project.urls]
Homepage = "https://github.com/SimplHDL"
Documentation = "https://simplhdl.readthedocs.io/en/latest/"
Repository = "https://github.com/SimplHDL/simplhdl.git"
Issues = "https://github.com/SimplHDL/simplhdl/issues"

[project.scripts]
simpl = "simplhdl.__main__:main"

[project.optional-dependencies]
dev = [
    "flake8",
    "mypy",
    "robotframework",
    "tox",
]
docs = [
    "furo",
    "Sphinx",
    "sphinx_pyproject",
]

[project.entry-points."simplhdl.plugins"]
chisel_generator = "simplhdl_chisel.chisel:ChiselGenerator"
encrypt = "simplhdl_encrypt.encrypt.encryptflow:EncryptFlow"
fusersoc_parser = "simplhdl_fusesocparser.fusesocparser:FuseSocParser"
lint = "simplhdl_lint.lint.lintflow:LintFlow"
modelsim = "simplhdl_modelsim.modelsim.modelsimflow:ModelSimFlow"
peakrdl_generator = "simplhdl_peakrdl.systemrdl:PeakRdlGenerator"
quartus = "simplhdl_quartus.quartusflow:QuartusFlow"
quartus_generator = "simplhdl_quartus.spd:QuartusGenerator"
quartus-dse = "simplhdl_quartus.quartusdse.quartusdseflow:QuartusDseFlow"
quartus-export = "simplhdl_quartus.quartusexport.quartusexportflow:QuartusExportFlow"
questasim = "simplhdl_questasim.questasim.questasimflow:QuestaSimFlow"
rivierapro = "simplhdl_rivierapro.rivierapro.rivieraproflow:RivieraProFlow"
simplhdl_parser = "simplhdl_parser.simplhdlparser:SimplHdlParser"
vcs = "simplhdl_vcs.vcs.vcsflow:VcsFlow"
vivado = "simplhdl_vivado.vivadoflow:VivadoFlow"
vivado_generator = "simplhdl_vivado.ipxact:VivadoGenerator"
xsim = "simplhdl_vivado.xsim.xsimflow:XsimFlow"

[tool.setuptools.dynamic]
version = {attr = "simplhdl.__version__"}

[tool.setuptools.package-data]
"*" = ["*.j2"]

[tool.flake8]
max-line-length = 120
max_complexity = 10
per-file-ignores = [
    "__init__.py:F401",
]
